Moved AWSAppleSignIn to a separate SDK
