+++ AWSAuthSDK/Sources/AWSAppleSignIn/AWSAppleSignInButton.h
