+++ AWSAuthSDK/Sources/AWSAppleSignIn/AWSAppleSignInButton.m
