+++ AWSAuthSDK/Sources/AWSAppleSignIn/AWSAppleSignInProvider.h
