+++ AWSAuthSDK/Sources/AWSAppleSignIn/AWSAppleSignInProvider.m
